Add base framework for functional configuration of Client.  (#5)
## What changes are proposed in this pull request?

This PR adds the base framework for functional configuration of modular
API Clients (either REST or gRPC).

The principle behind that framework is to build a strong abstraction
between the configuration surface (i.e. the functional options) and the
actual config (which is internal). This allows to easily evolve and
extend the configuration surface without introducing breaking changes.

**Why functional options and not method on a struct with private
members?** It is more idiomatic and modular while ensuring that each
clients has its own immutable config.

## How is this tested?

No test yet, this PR is purely meant as a "demo" to be extended in an
actual implementation (e.g. by implementing a specific roundtripper for
the HTTP client).