Add dummy Jobs v2 client. (#11)
## What changes are proposed in this pull request?

This PR adds a dummy Jobs v2 client. The code is meant as a way to test
current assumption and packages while providing an example for future
code generation.

Design decisions:

- `unstable` package: Having `ClientOptions` in `internal` provides
great abstraction, preventing SDK client for accessing and modifying the
internal client config representation. However, it also prevents the API
client to access the logger that was already resolved when building the
HTTP client. Package `unstable` provide a small utility to resolve the
logger from a set of options. It works for now but we might want to
consider alternatives (e.g. have `ClientOptions` in `unstable`).
- `genhelpers.go`: Most generated code is expected to go through the
same logic to send a request, log request/response, and parse errors.
The `genhelpers.go` file is meant to be a static file containing all
these helpers.
- Lazy loggers: The new `slog` package provides utility to lazily
generate logging entry. We will leverage these to log request/response
if the logger is configured at the right level.

## How is this tested?

No added test.