Add `RetryOnCodes` to easily construct `Retriers` that retry on specific error codes. (#10)
## What changes are proposed in this pull request?

This PR uncomment the `RetryOnCodes` function now that `apierr` has been
merged. The goal of this function is to provide a convenient way for
users to configure retry behavior based on specific API error codes
(e.g. retry only on Internal errors). The `WithDisableRetry()` option
complements the retry functionality by giving users an easy way to opt
out of automatic retries when needed.

Note about performance: one might be wondering why the implementation is
not using a set/map to check whether the error code is retriable or not.
The reason is that we expect the list of retriable codes to be small so
that iterating on the slice and comparing the integer values is actually
faster than a `_, ok := map[c]` check. This can easily be adjusted
transparently in the future.

## How is this tested?

Added unit tests for the new retrier.